Move get_response_status() over to ProfileErrors method
